requirements.txt
selenium
pillow
pywin32
